# docs/swagger/swagger.yaml
definitions:
...
  handlers.addParticipantRequest:
    properties:
      role:
        enum:
        - participant
        - responsible
        example: participant
        type: string
      user_id:
        example: 5
        type: integer
    type: object
...
  handlers.createTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      due_date:
        example: "2026-05-10T18:00:00Z"
        type: string
      status:
        enum:
        - todo
        - in_progress
        - done
        example: todo
        type: string
      title:
        example: Подготовить макеты
        type: string
    type: object
...
  handlers.participantResponse:
    properties:
      created_at:
        type: string
      event_id:
        type: integer
      id:
        type: integer
      role:
        type: string
      rsvp_status:
        type: string
      user_id:
        type: integer
    type: object
...
  handlers.taskResponse:
    properties:
      assignee_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      event_id:
        type: integer
      id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  handlers.updateParticipantRequest:
    properties:
      role:
        enum:
        - participant
        - responsible
        type: string
    type: object
  handlers.updateRSVPRequest:
    properties:
      rsvp_status:
        enum:
        - pending
        - accepted
        - declined
        type: string
    type: object
  handlers.updateTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      status:
        enum:
        - todo
        - in_progress
        - done
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: Сервис для управления рабочими мероприятиями.
  title: Meeting Service API
  version: "1.0"
paths:
...
  /events/{id}/participants:
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.participantResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Список участников встречи
      tags:
      - participants
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Пользователь и роль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.addParticipantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.participantResponse'
        "400":
          description: validation error
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Добавить участника во встречу
      tags:
      - participants
  /events/{id}/participants/{participantId}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Participant ID
        in: path
        name: participantId
        required: true
        type: integer
      responses:
        "204":
          description: no content
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: participant not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить участника из встречи
      tags:
      - participants
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Participant ID
        in: path
        name: participantId
        required: true
        type: integer
      - description: Новая роль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.updateParticipantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.participantResponse'
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: participant not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Изменить роль участника
      tags:
      - participants
  /events/{id}/participants/{participantId}/rsvp:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Participant ID
        in: path
        name: participantId
        required: true
        type: integer
      - description: Статус RSVP
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.updateRSVPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.participantResponse'
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: participant not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Ответить на приглашение (RSVP)
      tags:
      - participants
  /events/{id}/tasks:
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.taskResponse'
            type: array
      security:
      - BearerAuth: []
      summary: Все задачи встречи
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Параметры задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.createTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.taskResponse'
        "400":
          description: validation error
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создать задачу во встрече
      tags:
      - tasks
  /events/{id}/tasks/{taskId}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      responses:
        "204":
          description: no content
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: task not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удалить задачу
      tags:
      - tasks
    get:
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.taskResponse'
        "404":
          description: task not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получить задачу по id
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: Новые поля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.updateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.taskResponse'
        "400":
          description: validation error
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: task not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновить задачу
      tags:
      - tasks
...
swagger: "2.0"